VibeCodeCLI/requirements/base.txt
# Base requirements for AI Coding Agent
textual>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
aiohttp>=3.8.0
pydantic>=2.0.0
click>=8.0.0
pyperclip>=1.8.0
